Adding an example moveset
This is an example of a moveset that would be used in DMN's mod. I created this particular moveset based off my own personal one, simplified down a bit, and added in instructions to try and help new players to figure out what's going on. It is not refined in any way if i recall, as I was intending to have another crack at making it better, but it is a start.